Balance Level up

+++ DSVJ3_GAME/Assets/Scripts/Autobattle/Characters/Stats.cs
@@ -3,6 +3,6 @@ public struct Stats
 {
     public int damage;
     public int maxHealth;
-    public int attackSpeed;
-    public int armor;
+    public float attackSpeed;
+    public float armor;
 }
